services/api-gateway/requirements.txt
# =========================================================
# Web framework
# =========================================================
fastapi>=0.110
uvicorn[standard]>=0.27


# =========================================================
# Configuration & validation
# =========================================================
pydantic[email]>=2.5
pydantic-settings>=2.2
pydantic


# =========================================================
# Authentication & security
# =========================================================
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7.4
bcrypt<4.1


# =========================================================
# Database
# =========================================================
sqlalchemy>=2.0
asyncpg>=0.29          # if using PostgreSQL async
psycopg2-binary>=2.9  # if using PostgreSQL sync (optional)


# =========================================================
# Observability & metrics
# =========================================================
prometheus-client>=0.19


# =========================================================
# Utilities (safe defaults)
# =========================================================
python-dotenv>=1.0
